Fix: Avoid drag-drop above basis key
As mentioned in !141930, prevent shape key from dropping above the
basis, so it remains the first.

+++ source/blender/editors/object/interface_template_shape_key_tree.cc
@@ -98,6 +98,7 @@ class ShapeKeyDropTarget : public ui::TreeViewItemDropTarget {
     if (drag_shapekey->index == drop_index_) {
       return false;
     }
+
     return true;
   }
 
@@ -112,6 +113,9 @@ class ShapeKeyDropTarget : public ui::TreeViewItemDropTarget {
         BLI_assert_unreachable();
         break;
       case ui::DropLocation::Before:
+        if (drop_index_ == 0) {
+          return TIP_("Cannot move above basis shape key");
+        }
         return fmt::format(fmt::runtime(TIP_("Move {} above {}")), drag_name, drop_name);
       case ui::DropLocation::After:
         return fmt::format(fmt::runtime(TIP_("Move {} below {}")), drag_name, drop_name);
@@ -134,6 +138,9 @@ class ShapeKeyDropTarget : public ui::TreeViewItemDropTarget {
         BLI_assert_unreachable();
         break;
       case ui::DropLocation::Before:
+        if (drop_index == 0) {
+          return false;
+        }
         drop_index -= int(drag_index < drop_index);
         break;
       case ui::DropLocation::After:
